refactor: useWidgetHook usage
